Some small performance tweaks
This should be good enough for about ~35 speed up; 11.8 vs 17.1s on

```
data(fakeOutbreak)
dat <- fakeOutbreak$dat
w <- fakeOutbreak$w

set.seed(1)
system.time(
out <- outbreaker(data=list(dna=dat$dna, dates=dat$onset, w.dens=w),
                   config=list(n.iter=100, sample.every=5)))
```

However, they are not extensively tested and this does disable a bunch
of testing.  I think that the testing is in the wrong place - we still
need to do it, but not in the main loop as is currently done.