fix new_kappa allocation
With the previous script, the allocation of new_kappa using param["kappa"] was interfering with the pointer definition. It was then causing errors in the computation of the new and old likelihoods (new_likelihood then corresponded to the old_likelihood).
In outbreaker_find_imports: there was a mistake in the position of the parentheses for the calculation of n_measures.